+++ DigitRecognitionNN/Models/Matrix.cs
@@ -4,7 +4,7 @@ namespace DigitRecognitionNN.Models;
 
 public class Matrix
 {
-    private readonly double[,] data;
+    private readonly float[,] data;
     public int Rows { get; }
     public int Cols { get; }
     
@@ -13,10 +13,10 @@ public class Matrix
     {
         Rows = rows;
         Cols = cols;
-        data = new double[rows, cols];
+        data = new float[rows, cols];
     }
 
-    public Matrix(double[,] data)
+    public Matrix(float[,] data)
     {
         this.data = data;
         Rows = data.GetLength(0);
@@ -66,7 +66,7 @@ public class Matrix
         {
             for (int j = 0; j < b.Cols; j++)
             {
-                double sum = 0;
+                float sum = 0;
                 for (int k = 0; k < a.Cols; k++)
                 {
                     sum += a[i, k] * b[k, j];
@@ -77,7 +77,7 @@ public class Matrix
         return result;
     }
 
-    public static Matrix operator *(Matrix a, double b)
+    public static Matrix operator *(Matrix a, float b)
     {
         var result = new Matrix(a.Rows, a.Cols);
         for (int i = 0; i < a.Rows; i++)
@@ -128,7 +128,7 @@ public class Matrix
     //public void SetZero() => Array.Clear(data, 0, data.Length);
     
     // Indexer
-    public double this[int row, int col]
+    public float this[int row, int col]
     {
         get => data[row, col];
         set => data[row, col] = value;
@@ -136,9 +136,9 @@ public class Matrix
 
     
     // Transform
-    public double[] ToArray()
+    public float[] ToArray()
     {
-        double[] result = new double[Rows * Cols];
+        float[] result = new float[Rows * Cols];
         int k = 0;
         
         for (int i = 0; i < Rows; i++)
@@ -153,7 +153,7 @@ public class Matrix
         return result;
     }
 
-    public static Matrix FromArray(double[] array)
+    public static Matrix FromArray(float[] array)
     {
         var result = new Matrix(array.Length, 1);
         for (int i = 0; i < array.Length; i++)
@@ -163,18 +163,18 @@ public class Matrix
         return result;
     }
 
-    public double[][] ToJaggedArray()
+    public float[][] ToJaggedArray()
     {
-        double[][] result = new double[Rows][];
+        float[][] result = new float[Rows][];
         for (int i = 0; i < Rows; i++)
         {
-            result[i] = new double[Cols];
+            result[i] = new float[Cols];
             for (int j = 0; j < Cols; j++)
                 result[i][j] = this[i, j];
         }
         return result;
     }
-    public static Matrix FromJaggedArray(double[][]? array)
+    public static Matrix FromJaggedArray(float[][]? array)
     {
         if (array == null)
             throw new ArgumentNullException(nameof(array), "Array cannot be null.");
